+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## [Unreleased]
 
 ### Added
+- Handle `important` field property.
+- Handle `underscoreUpper` for names (Example: FIRST_NAME).
+- Handle `hyphenUpper` for names (Example: FIRST-NAME).
+- Handle subtype `rich` of type `string`.
 
 ### Changed